Train Step Compile [2/3] optimizer support
Support compiling a user-provided train_step function, provided
- it includes exactly one call to .backward()
- it includes any relevant optimizer steps, followed by zero_grad() calls

Make Adam optimizer (almost) work

Currently, use faketensor execution for the first optimizer step,
to create correct-shaped fake tensors for optimizer states, and
avoid taking the init path on the second try.

Then we hackily convert the optimizer states to zero tensors.  This is the
worst part of the current impl, as it fails to trace the real optimizer
init and could be incorrect.

Finally, we can wrap the real optimizer state tensors with fake tensors
to perform a real train step trace.

ghstack-source-id: 1a52d76df44a4009b8af6b5f913aae7769bc529b
Pull Request resolved: https://github.com/pytorch/pytorch/pull/98461